modify uint64_t print

+++ ML/Common/thread.cpp
@@ -33,7 +33,7 @@ int Thread::start() {
 
 void Thread::join() {
 	if (_tid > 0) {
-		LOG_DEBUG("tid : %lu", _tid);
+		LOG_DEBUG("tid : %llu", (uint64_t)_tid);
 		pthread_join(_tid, NULL);
 	}
 }
